@@ -3,25 +3,44 @@
   "version": "1.0.0",
   "description": "",
   "main": "gulpfile.js",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "gulp watch",
+    "build": "CI=true gulp",
+    "lint": "CI=true gulp lint"
   },
-  "author": "",
-  "license": "ISC",
+  "engines": {
+    "node": ">=18.20.0 <=20.0.0"
+  },
   "devDependencies": {
-    "eslint": "^8.2.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.29.1",
-    "gulp": "^4.0.2",
-    "gulp-concat": "^2.6.1",
+    "autoprefixer": "^10.4.19",
+    "breakpoint-sass": "^3.0.0",
+    "browser-sync": "^3.0.2",
+    "del": "^7.1.0",
+    "eslint": "^8.9.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^3.4.1",
+    "gulp": "^5.0.0",
+    "gulp-eslint-new": "^2.0.0",
+    "gulp-if": "^3.0.0",
+    "gulp-imagemin": "^9.1.0",
+    "gulp-postcss": "^10.0.0",
+    "gulp-prettier": "^5.0.0",
     "gulp-sass": "^5.1.0",
     "gulp-sass-glob": "^1.1.0",
+    "gulp-stylelint": "^13.0.0",
     "gulp-sourcemaps": "^3.0.0",
-    "gulp-uglify": "^3.0.2"
-  },
-  "dependencies": {
-    "del": "^6.1.1",
-    "gulp-eslint": "^6.0.0",
-    "sass": "^1.77.2"
+    "gulp-svgstore": "^9.0.0",
+    "node-sass-globbing": "^0.0.23",
+    "postcss": "^8.4.38",
+    "postcss-scss": "^4.0.9",
+    "prettier": "^2.7.1",
+    "sass": "^1.77.0",
+    "stylelint": "^13.13.1",
+    "stylelint-config-prettier": "^9.0.3",
+    "stylelint-prettier": "^1.2.0",
+    "stylelint-scss": "^3.21.0"
   }
 }
